+++ order-service/src/main/java/com/eigtheenthstreet/order_service/exception/company/CustomCompanyGetApiFailException.java
@@ -1,0 +1,11 @@
+package com.eigtheenthstreet.order_service.exception.company;
+
+import com.eigtheenthstreet.order_service.exception.CustomException;
+
+import exception.ErrorCode;
+
+public class CustomCompanyGetApiFailException extends CustomException {
+	public CustomCompanyGetApiFailException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
