+++ order-service/src/main/java/com/eigtheenthstreet/order_service/exception/order/CustomOrderStatusUpdateNotAllowedException.java
@@ -1,4 +1,6 @@
-package com.eigtheenthstreet.order_service.exception;
+package com.eigtheenthstreet.order_service.exception.order;
+
+import com.eigtheenthstreet.order_service.exception.CustomException;
 
 import exception.ErrorCode;
