+++ order-service/src/main/java/com/eigtheenthstreet/order_service/exception/product/CustomProductBulkNotFoundException.java
@@ -1,0 +1,12 @@
+package com.eigtheenthstreet.order_service.exception.product;
+
+import com.eigtheenthstreet.order_service.exception.CustomException;
+
+import exception.ErrorCode;
+
+public class CustomProductBulkNotFoundException extends CustomException {
+
+	public CustomProductBulkNotFoundException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
