+++ order-service/src/main/java/com/eigtheenthstreet/order_service/exception/product/CustomProductDecreaseStockApiFailException.java
@@ -1,0 +1,11 @@
+package com.eigtheenthstreet.order_service.exception.product;
+
+import com.eigtheenthstreet.order_service.exception.CustomException;
+
+import exception.ErrorCode;
+
+public class CustomProductDecreaseStockApiFailException extends CustomException {
+	public CustomProductDecreaseStockApiFailException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
